Fixes #1009: Race condition with network thread in ElectrumWindow c'tor
In ElectrumWindow.__init__, the network callbacks were being registered
before self.wallet was defined.  The callbacks rely on self.wallet being
defined for them to operate correctly.  Since the callbacks run in
another thread, it was possible for an exception to be thrown as
self.wallet was not defined. Rare, but was a race condition. See #1009.