[Network] Fix #956 - unverified dict modified while iterating
The unverified dict from the wallet was being iterated over in the
verifier -- but it could lead to a situation where it would end up
modifying itself while iterating (SPV.verify_merkle() specifically).

This has now easily been fixed by iterating over a copy() of the dict,
rather than the dict itself. The dict is usually not very large so the
performance impact of this is minimal and the benefit is correct code.