[iOS] Fix #1072 -- Make sure there is enough disk space.
Previously a low-level OS error would occur if the app ran out of disk
space.  This could lead to bizarre/unexpected errors and even a crash
reporter in some cases.

So I added a check for sufficient disk space every time the app is
foregrounded, and if <50MB space, stop the daemon and tell the user that
Electron Cash cannot proceed.

This fixes #1072.