[Qt] Fix #953 -- Esoteric C++ crash
If the context menu in the address list was up over the 'label' column
(description) -- and the widget was refreshed, and then the user selected
"edit label" from the context label, the lambda internally would be
pointing to a QTreeWidgetItem that no longer existed on the C++ side,
and thus cause a RuntimeError.

This has been fixed by not storing the QTreeWidgetItem associated with
the context menu in a lambda's cell, but instead re-computing it in the
lambda from right-click position again.

It's hard to explain.  Read the commit code to understand. :)