[Qt] Fix #921 -- InvalidPassword on private key export
I suspect this could only happen if the password changes while exporting
or there is some wallet corruption.

Regardless -- we now catch the exception and put "INVALID_PASSWORD" in
the exported data.

I couldn't actually reproduce this error -- but fixing the bug anyway in
case wallet corruption is the culprit. Rationale is user is better off
at least being able to export SOME of his keys rather than NONE in case
of corruption.