[Lib] Synchronizer was enqueueing empty subscribe-to-scripthashes
Every 100ms the Synchronizer for an HD wallet was enqueuing an empty
"subscribe to addresses" call to the network layer.  This didn't
actually do anything and wouldn't enqueue anything on the network. All
it would do is cause a lock to be acquired and released twice needlessly
every 100ms, wasting previous CPU cycles.

I added code to avoid this situation.