[Lib] Fix #985. OverflowError if user specified huge port number
See #985.  It was possible for users to specify a gigantic port value in
the UI and this would be taken verbatim as a python int (python ints
have infinite range) and passed along to C ultimately in the socket
code. Of course, C ints don't have an unlimited range and this was
raising an OverFlow error at the Python <-> C boundary layer.

Fixed to catch this exception and treat it like any other invalid socket
parameter -- calling code just connect/create the socket (which is what
you'd expcet rather than a crash).