[Cosigner] Finally addresses issue #1113
Since the cosigner window can arrive at any time (including while user
is typing/hitting enter or space), we want the 'default' button for the
cosigner window to be as harmless as possible.

It was not possible to NOT have a default button due to Qt quirks.  So
instead, the cosigner popup has a "Show details..." button which just
reveals some help text explaining what's going on.

This way, if the window arrives -- and the user is typing -- nothing
destructive can really happen.