[Qt][Lib] QR codes now are now mixed-case.
Issue reported by @zquestz:

BITCOINCASH: upper-case QR codes, while in the spec, are not actually
implemented on all wallets.  As a result, the QR codes generated by
Electron Cash were sometimes invalid in wallets such as Edge and others.

We originally preferred uppercase for the address QR codes because they
generate small/clearer QR images -- but this benefit is not enough to
justify breaking compatibility with broken wallets.