[Qt] Added TxUpdateMgr class to manage 'verifid2' and 'new_transaction'
Put the code that collates 'verified2' and 'new_transaction' network
calls into its own class, thus making class ElectrumWindow easier to
maintain.  The class is called TxUpdateMgr, and it lives in
main_window.py but it encapsulates all the logic for collating and
updating the gui for 'verified2' and 'new_transaction' network callbacks
in an efficient manner. This mechanism was introduced in a recent commit
but the code for it was all over the place.  Putting it in 1 class makes
everything more maintainable.