[Lib] Synchronizer fix to allow the server to not find a tx
From Electrum commit 7b8114f865f644c5611c3bb849c4f4fc6ce9e376
Electrum issue #5122

Adapted to our codebase, however, so it's not a cherry-pick

SomberNight said:
"User has wallet file with history that includes some txid; corresponding
raw tx is not in the "transactions" dict in the file however.
When the synchronizer starts up, it requests this "missing" txn from
the server... but what if the server does not know about it?
Maybe it was reorged and is not in the new best chain,
and not even in mempool. This was not handled previously."