Import all instances of decimal.Decimal as PyDecimal in codebase
This is a paranoia fix in reference to bug #1157. Qt 5.12 also exports
the 'Decimal' symbol if you import *, and since Qt's imports sometimes
appears *after* the python imports, it may win.

We are renaming Python decimal.Decimal to PyDecimal in the codebase to
allow code to also access the Qt Decimal.

Also note that external plugins may be affected.  It is hoped that by
explicitly renaming the Python Decimal -> PyDecmal in the import, the
likelihood of a plugin being hit by this is minimal.

We needed to also rename in the lib/ directory (which doesn't normally
see the Qt namespace) because some Qt-based plugins import * from lib/
as well as from Qt.