[iOS] Updated to Python 3.6.6, plus small tweaks
- Updated the 'make_ios_project' system to use Python 3.6+.  Python 3.5
is going to be phased out, but still unofficially works.

- iOS app has a much shorter reconnect interval on connection down; this
is necessary because mobile connections are quite spotty.

- iOS app has a hardcoded "preferred server only" setting internally for
now.  This is a security measure.  Users were getting intermittent
problems due to the phishers so we block them altogether on iOS.

- Refactored some stuff in network.py to allow iOS to specify its own
reconnect interval more easily.

- On iOS we always display the TOTAL balance (confirmed + unconfirmed
added together).  This is less confusing to newbie users and on BCH with
our reliable 0-conf, unconfirmed is not such a big deal.  Unconfirmed
balances are highlighted in BLUE to alert users that it's not confirmed
yet, but otherwise the UI shows the total balance (conf + unconf).