[Lib] Fixups to bitcoin.py deserialize_privkey etc w.r.t bip38
Before the code was bad and would fail with a ValueError that callers
weren't expecting if in the install wizard user pasted a bip38 key.
These fixups ensure bip38 keys are rejected internally properly and
don't identify as WIF keys in the proper functions so that user doesn't
get a crash.

We will handle bip38 as a special case in the UI...